+++ CHANGES.md
@@ -2,6 +2,7 @@
 * simplify some `after_theme_row`
 * better internationalization for changing plugin _View details_ link
 * update `after_theme_row` classes for Shiny Updates
+* refactor and improve for `GitHub Updater Additions` plugin
 
 #### 5.4.1
 * get tags for themes to rollback even if no updates are available. I was overzealous in cutting remote API calls.